refactor : TSK-17 github, clova 타입 분리
